Always compare and sync status on hub (#67)
Policy status update will be skipped for hub if the status is not
changed on the managed cluster. This won't work when a cluster
is resumed from hibernation where the status on managed cluster
is not changed however the status on hub was updated due to
hibernation.

For https://github.com/open-cluster-management/backlog/issues/17867

Signed-off-by: Yu Cao <ycao@redhat.com>